arm64: dts: ADRV9002 ZCU102 FFH Updates
Updated the FFH device tree to refernce the Rx2Tx2 based dts file.
Changed the FFH TDD clock to match the updated HDL, extended the
calibration range for FFH, and corrected a typo with the FFH
mode.

Added a device tree for using an external ADF4351 as the ADRV9002
clock source.

Dummy hack to the ADF4350/1 driver to comply with the clock
subsystem.

Signed-off-by: Brent Kowal <brent.kowal@analog.com>